Separate renderer logic into three new files
Took 1 hour 21 minutes
